0.81.0 [BIG BUG FIXED!] close and reopen GUI will reset all sliders values!
THIS BUG WAS FIXED! OMG I CAN'T BELIEVE IT STARTED FROM VERSION 0.706!!!!!
close and then reopen GUI, all GUI will recreated, and comboBoxChanged will be called, so it will set values to preset value. If you choose a preset and then make some changes on that preset, close and reopen the plugin will set values to preset default!
But thank God, this was finally fixed!


Former-commit-id: 6edbeace9038e352ec5da6da6e713349cac72bcd