fix(dsp): Correct LFO bipolar modulation range
The LFO modulation in bipolar mode had a range twice as large as in unipolar mode for the same depth value. This led to counter-intuitive behavior, especially when modulating a parameter from its minimum value, where it would unexpectedly span the entire parameter range instead of half.

This commit also changes UI.